port: FILE type bridge — 21 stdio functions call real platform.posix
Changed FILE.handle from Int to Long (64-bit pointer).
21 stdio functions now call real C via platform.posix:
fopen, freopen, fflush, fclose, tmpfile, fgetc, fputc, fputs,
ungetc, fseek, ftell, feof, ferror, rewind, fdopen, perror,
remove, rename, getchar, putchar, puts.

6 functions still throw (setvbuf, setbuf, fread, fwrite, fgetpos,
fsetpos) — need CValuesRef/FposT bridge for buffer/position params.

All tests pass.

+++ src/commonMain/kotlin/io/github/kotlinmania/libc/COpaqueTypes.kt
@@ -29,7 +29,7 @@ public class COpaquePointer(
  * On JS/WASM, [handle] holds the numeric FILE pointer from N-API.
  */
 public class FILE(
-    public val handle: Int = -1,
+    public val handle: Long = -1L,
 )
 
 /**
